ES6-ized some code!
Comments are best placed with a line break before them (unless the
comment is the first line of code in a function) and functions with the
opening bracket on a new line.

I added some ES6 code to make this more elegant and less complex! :)

Also it’s best to use !== and === as opposed to != and == in order to
ensure all code matches the data type you expect, that can lead to pain
in the future! :)