Some housecleaning
It’s best to have single-line td elements with zero line spaces between
them and block ones (with child elements) having a line break between
them and afterwards (unless it’s the final td).

The bicolour attribute is deprecated in HTML, please use CSS or inline
styling for that.

Please use spinal-case variables for CSS classes, e.g. table-headers

Also strictly camelCase for variables, in this case k_value should be
kValue.

This may sound like a lot but you’ll get used to this in no time at all
and hopefully enjoy the look of the code and find it easier to work
with! :)

When inline styling is used it’s best to have a space after the first
colon as with a few styles it can quickly become messy :)

Finally I fixed the code hierarchy on the each loop.